Put all excemption in a file
Not sure if this solves anything. Consider simply writing a test that runs quickjs on fake github fs
